refactor(mtd): update NOR bank management
All the quad or octal SPI NOR memories do not support bank selection.
Warn the user in case of the memory exceeds 16 MB and the read opcode
address is set to 3 bytes and the bank selection flag has not been set.

This patch enables the bank selection flag on MP1 board.

Change-Id: Id3004ccc865b7888b891140da2c0335eb1dbbc16
Signed-off-by: Christophe Kerello <christophe.kerello@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/248751
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Christophe KERELLO <christophe.kerello@st.com>
Reviewed-by: Yann GAUTIER <yann.gautier@foss.st.com>
Tested-by: Christophe KERELLO <christophe.kerello@st.com>