fix(st-ddr): manage AXI ports and HIF when updating QD3 registers
Quasi-dynamic group 3 (QD3) registers are updated when DDR
is not completely initialized, i.e. when AXI ports are not enabled and
possibly when host interface (HIF) is not enabled too.
In that case, a specific mechanism is necessary to restore the same
conditions as before accessing QD3 registers.
Static functions have been added to get AXI ports and HIF states and
are used to determine the needed conditions to set/unset.

Signed-off-by: Nicolas Le Bayon <nicolas.le.bayon@st.com>
Change-Id: I820c7fc04981dbd978696b5c00e75820aa392907