feat(stm32mp2): reserve the first 4KB of SRAM1 for BSEC3 driver
The STM32MP25 OTP values and status are shadowed in the first 4KB
of SRAM1 by the BSEC3 driver.

Reserve this memory in SRAM1 for BL2 and in BL31 to prepare the
support of this region in BSEC3 driver.

When STM32MP_DDR_FIP_IO_STORAGE is activated all the SRAM1 is reserved
for BL2 with MAP_SRAM1 region.

Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: Ie9204022866d531f8c2c125efdffc2c8e1b42891
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/258478
Tested-by: Yann GAUTIER <yann.gautier@foss.st.com>
Reviewed-by: Yann GAUTIER <yann.gautier@foss.st.com>
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Nicolas LE BAYON <nicolas.le.bayon@st.com>
Domain-Review: Yann GAUTIER <yann.gautier@foss.st.com>