feat(st): add an API to retrieve STM32 image header
Add an helper function stm32_get_header_address() to be used
in get_rotpk_hash().
For the moment it just returns STM32MP_HEADER_BASE.
This macro is then defined for STM32MP13 and STM32MP15.
The macro is not directly used in get_rotpk_hash(), as the header
address will change depending on the use case.

Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Change-Id: Ib78a8cfac9b7514229d654e4d71651d7c33ac5dc