fix(st): fix uart issue in SSP mode
NOT_UPSTREAMABLE

Add a flush to ensure that the programmer get time to read the last
command sent.
Remove the is_valid_header that is only used in case of FIP. This
must not be checked in case of SSP mode.

Signed-off-by: Lionel Debieve <lionel.debieve@foss.st.com>
Change-Id: I456d996d00e48abc712c50556e414ff4e5552466
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/294990
Reviewed-by: Yann GAUTIER <yann.gautier@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>