feat(stm32mp1): check the backup ram access
In case of boot when a tamper is raised, the backup RAM access
can be forbidden and a control access need to be done to avoid
abort.
In that case, we will notice that the BL2 is booting in a degraded
mode because no suspend context can be saved.

Signed-off-by: Lionel Debieve <lionel.debieve@foss.st.com>
Signed-off-by: Maxime Méré <maxime.mere@foss.st.com>
Change-Id: I9063b2da3d4dbe3f0c458171dd7aef9c34d74d51
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/333573
Reviewed-by: Yann GAUTIER <yann.gautier@foss.st.com>
Domain-Review: Yann GAUTIER <yann.gautier@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>