feat(stm32mp2): add function plat_get_target_pwr_state
Add plat_get_target_pwr_state for STM32MP25 platform as the default
implementation in plat_spci_common is not working.

The power state is RUN = 0U for simple CPU STANDBY, with fast path
in psci_cpu_suspend(), so the function will ALWAYS return 0U
for coordination with 2 cores for STM32MP25.

Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: I35efd5d5b161ac4e406190985e83f97342dac7db