fix(stm32mp2): keep ECSDA key and signature definitions
The patch [1] was removing BOOT_API_ECDSA_PUB_KEY_LEN_IN_BYTES and
BOOT_API_ECDSA_SIGNATURE_LEN_IN_BYTES, and defining 256 and 384
versions of the macros. But the removed macros are required in
crypto lib when compiling with trusted boot. Redefine them using
the 256/384 version depending on SoC is STM32MP1 or not.

[1] 4cf19ad4de feat(stm32mp21): update boot API header version

Change-Id: I4ec8e8b9d249ed542dda9bd64ef6248f8a56ef62
Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/365744
Reviewed-by: Maxime MERE <maxime.mere@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Nicolas LE BAYON <nicolas.le.bayon@st.com>