feat(stm32mp2): optimize the wake-up time for low power modes
Align the PWR and RCC delay and configuration for PWR pins (PWR_ON,
PWR_CPU_ON and PWR_LP) with the default values defined in AN5727 for
STM32MP25x boards with STPMIC25.

This patch allows the configuration of these low power parameters with
device tree, if the default parameters defined for STM32MP25 SoC with
STPMIC25 are not correct for the customer board.

Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: I0c60af36de7681552b61d8a6d4e168be422da575
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/358588
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/367563
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/367815
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/368671
Reviewed-by: Yann GAUTIER <yann.gautier@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>