feat(stm32mp1): rework ddr supplies handling
Retrieve DDR power supplies from the device-tree.
This patch requests to describe the DDR power supplies
in the DT_DDR_COMPAT node.
This patch assumes also that the sink-source or
the bypass property was set in the device-tree in the ldo3
node.

Change-Id: Ida9dd705b623a8bd40dd206cc08e2d550446b172
Signed-off-by: Pascal Paillet <p.paillet@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/388115
Domain-Review: Patrick DELAUNAY <patrick.delaunay@foss.st.com>
Reviewed-by: Yann GAUTIER <yann.gautier@foss.st.com>
Tested-by: Pascal PAILLET <p.paillet@st.com>
Reviewed-by: Pascal PAILLET <p.paillet@st.com>