feat(st-clk): move oscillators ops from clk-stm32-core to drivers
In the context of M33_TDCID, TF-A cannot enable/disable oscillators.
The ops should then be under flag #if !STM32MP_M33_TDCID.
This is OK for the MP2 clock driver, as the flag is always set
(to 0 by default), but the flag doesn't exist on MP13, which uses
the same clock framework.
To avoid this the oscillators ops structure should be move to the driver.

Signed-off-by: Gabriel Fernandez <gabriel.fernandez@foss.st.com>
Change-Id: Ibdef015119a17cc50a151f7e3a11d168376d11ca
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/393869
Domain-Review: Yann GAUTIER <yann.gautier@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Maxime MERE <maxime.mere@foss.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>