fix(stm32mp2): stm32_otp_is_closed_device should return false
If Cortex-M33 is TDCID we cannot know the status of BSEC. The function
stm32mp_check_closed_device() can then not rely on this to know if
the device is closed. We can only rely on lifecycle OTP. Change the
return of stm32_otp_is_closed_device() used when M33 is TDCID to
false, to be able to check lifecycle OTP value.

Change-Id: Ia16a5744c3065fed5039e630cde725fd88d719ce
Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/398289
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Maxime MERE <maxime.mere@foss.st.com>