refactor(stm32mp2): update authentication check for STM32MP2
The platform allow authentication check when chip is open to comfirm
that the key hash stored in OTP is correct. Put all that code directly
in authentication_check() function in order to simplify it.

In the STM32 MPU, we use auth_status from boot_api_context to display
the authentication check status on the console for CLOSED
UNSECURED/UNLOCKED samples. Similar to the MP25, the MP21 status
reflects only the signature check and not the ROTPK check.

However, unlike the STM32MP25, the STM32MP21 platform doesn't allow
reading the OTP that contains the ROTPK Hash, so there is no way to be
sure that the authentication succeeded when the chip is in an open
state. The authentication check code is then useless in the STM32MP21,
so the dedicated portion of code should be stubbed.

Signed-off-by: Maxime Méré <maxime.mere@foss.st.com>
Change-Id: Id20e813cdff751d000011fc33a55c90065076966
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/403011
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Domain-Review: Yann GAUTIER <yann.gautier@foss.st.com>
Reviewed-by: Yann GAUTIER <yann.gautier@foss.st.com>