Trace in stm32_get_sys_suspend_power_state
When Standby1 is limited by TF-A BL31, it is difficult to understood the
cause of the issue without modifing the code.

This patch allows to have the minimal information in console needed to
understood this restriction, when the console is not suspended by Linux
kernel in system suspend with:

echo N > /sys/module/printk/parameters/console_suspend

By default the console is suspended in suspend mode and the trace is not
displayed and the added latency for trace is acceptable for this use case
(vs time to freeze linux tasks).

Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: Ib9ea42d4cde3bb5592b86c5ddaf7a11a03a38613
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/407876
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Maxime MERE <maxime.mere@foss.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Gabriel FERNANDEZ <gabriel.fernandez@foss.st.com>