fix(stm32mp2): restore system for warmboot after D1 DStandby exit
After Standby1 request, when the request is not accepted by cortex
M33 for cortex-M33 TDCID, the D1 domain stay in DStandby mode and the
content of the SYSRAM is not lost.

On next wake-up, the TF-A BL31 warm boot entry is used by ROM code,
and TF-BL2 is not executed, as for Stop2 mode exit.

But in this case, initializations of the CA35 subsystem are missing
after cluster reset: GIC configuration and the generic timer
(based on STGEN).

The standby flag RCC_C1BOOTRSTSCLRR_STBYC1RSTF is tested because,
the GIC function stm32mp_gic_resume() can't not called for Standby1 exit
as  they based on configuration saved SYSRAM in stm32mp_gic_save(),
lost is this mode.

Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: Ic6bded2a4bca3719ad479cb17bab99433ec23023
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/429571
Reviewed-by: Maxime MERE <maxime.mere@foss.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>