feat(st-pmic): add default regulator configuration for enable_ramp_delay
Configure the default ramp delay for the STPMIC25, STPMIC2L and STPMIC1L
regulators, the DT configuration should be used only to change this
values, for example for board or external device constraints.

The behavior is aligned with STPMIC1 driver and the 1ms value is aligned with OP-TEE driver.
This patch allows to align the regulator part of the device tree
between OP-TEE and TF-A.

Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: I97f00fe118f89d45b3d7bc80ca11aa2e74d9e06c
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/443763
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Maxime MERE <maxime.mere@foss.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>