feat(st-crypto): add AES decrypt/auth by CRYP peripheral
Add code to be able to use STMicroelectronics CRYP peripheral. This
driver can manage many AES algorithms (CBC, ECB, CCM, GCM). It will be
used by the authenticated decryption framework (AES-GCM only).

Contrary to SAES, CRYP is not DPA-resistant, but it can use a shared key
with SAES, which is the safer way to use a secret key when A35 is not
TDCID and does not have direct access to OTP fuses. Most of this code is
based on OPTEE driver except for the key sharing part.

Change-Id: I7bb1e515026415428e108e19143e784cad497fb8
Signed-off-by: Maxime Méré <maxime.mere@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/466674
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Domain-Review: Yann GAUTIER <yann.gautier@foss.st.com>