feat(st): add interface to get a clock by an index from phandle
This adds an interface to retrieve a clock by its index from a phandle
in the device tree.

The phandle of the clock provider is stored in the upper 16 bits (MSB),
and the clock ID is stored in the lower 16 bits (LSB).
If #clock-cells = 0, the phandle is stored in both the MSB and LSB.

Signed-off-by: Gabriel Fernandez <gabriel.fernandez@foss.st.com>
Change-Id: I1dfbc666590ce648c588a0d3bc40ebb3d5f1243c
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/493936
Domain-Review: Patrick DELAUNAY <patrick.delaunay@foss.st.com>
Reviewed-by: Maxime MERE <maxime.mere@foss.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>